Adding all the project files in the repository.
I'm adding all the files in this repository after I've completed developing this project. As I keep track of my projects manually, I'm not very familiar with git and github. So I just developed it completely first, and am now adding the files on here so that the project can be saved on my profile.